Update Number to words to take a long

+++ src/Humanizer/Localisation/NumberToWords/FrenchNumberToWordsConverter.cs
@@ -4,7 +4,7 @@ namespace Humanizer.Localisation.NumberToWords
 {
     internal class FrenchNumberToWordsConverter : FrenchNumberToWordsConverterBase
     {
-        protected override void CollectPartsUnderAHundred(ICollection<string> parts, ref int number, GrammaticalGender gender, bool pluralize)
+        protected override void CollectPartsUnderAHundred(ICollection<string> parts, ref long number, GrammaticalGender gender, bool pluralize)
         {
             if (number == 71)
                 parts.Add("soixante et onze");
